backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
marshmallow==3.21.1
mysqlclient==2.2.4
python-dotenv==1.0.1
pdfplumber==0.11.9
alembic==1.13.1
